requirements.txt
flask==1.1.1
flasgger==0.9.1
flask-cors==3.0.8
psycopg2-binary==2.8.3
simplejson==3.16.0